Immuable class GeoPoint
